refactor: component detection optimisation to reduce function component related false positives
