+++ packages/ast/src/class/index.ts
@@ -1,0 +1,1 @@
+export * from "./class-identifier";
