+++ packages/ast/src/function/index.ts
@@ -1,0 +1,2 @@
+export * from "./function-identifier";
+export * from "./hierarchy";
