+++ packages/ast/src/index.ts
@@ -1,15 +1,11 @@
 export * from "./assignment";
 export * from "./call";
-export * from "./construction";
-export * from "./equal";
+export * from "./class";
 export * from "./eslint-community-eslint-utils";
-export * from "./hierarchy";
+export * from "./function";
 export * from "./identifier";
 export * from "./literal";
 export * from "./misc";
-export * from "./node-type";
-export * from "./property";
+export * from "./node";
 export * from "./return";
-export * from "./scope";
-export * from "./traverse";
 export * from "./variable";
