+++ packages/ast/src/literal/is.ts
@@ -1,7 +1,7 @@
 import { P } from "@eslint-react/tools";
 import type { TSESTree } from "@typescript-eslint/types";
 
-import { NodeType } from "./node-type";
+import { NodeType } from "../node";
 
 export function isRegExpLiteral(node: TSESTree.Node): node is TSESTree.RegExpLiteral {
   return node.type === NodeType.Literal && "regex" in node;
