--- packages/ast/src/node-type/index.ts
@@ -1,2 +1,0 @@
-export * from "./node-type";
-export * from "./readable-node-type";
