+++ packages/ast/src/node/equal.ts
