+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./api";
 export * from "./component";
+export * from "./construction";
 export * from "./context";
 export * from "./hook";
 export * from "./internal";
